выделение отрисовки полноразмерного кода в отдельный модуль
